// src/models/console-message.ts
export enum ConsoleMessage {
    TITLE = 'CGX',
    BANNER = 'Generate all the recommended markdown documentation/templates for the Github, Gitlab and Bitbucket community standards',
    ERROR = 'ERROR: ',
    SUCCESS = 'SUCCESS: ',
    INFO = 'INFO: ',
    GENERATE = 'GENERATE: ',
    CREATE = 'CREATE: ',
    UPDATE = 'UPDATE: ',
    START_GENERATING = 'Start generating all the recommended files...',
}